Resiliency benchmarking
I apologize for this being so big.  I didn't intend to redo the modules,
but it got big enough that I just went ahead and did it.  Also, while
it's big, this is sort of incomplete with respect to 'niceness' parity
with the performance work.  I'm unsure of priorities of this work and am
basically just getting this pushed off my box in case I get `rm -rf`
happy.  Anyhow, here are the highlights:

1. I used more module nesting so that my editor's file browser wasn't
one huge list.  I also hope the first-level names will help with project
comprehension.
1. I added the chaos work, which uses `saboteur` to inject network-based
chaos between the client and the server.
1. I added an explicit, Tomcat
server-closes-persistent-connection-due-to-timeout test (to some of the
module).
1. The mock server application was moved to Java8 for greater deployment
flexibility.  Basically, I was trying to use super old docker images in
order to get saboteur to work (most of that was futile).  This move
necessitated splitting out the `Payloads` class.
1. Reporting against the resiliency metrics resulted in fixing the
hideously coded, performance-specific reporting component.